requirements.txt
amqp==5.1.1
asgiref==3.7.2
astroid==2.15.6
billiard==4.1.0
black==23.7.0
celery==5.3.1
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.5
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
dill==0.3.6
Django==4.2.3
django-countries==7.5.1
django-debug-toolbar==4.1.0
django-extensions==3.2.3
django-filter==23.2
django-stubs==4.2.3
django-stubs-ext==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.6
flake8==6.0.0
flower==2.0.0
humanize==4.7.0
idna==3.4
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
kombu==5.3.1
lazy-object-proxy==1.9.0
MarkupSafe==2.1.3
mccabe==0.7.0
mypy==1.4.1
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.1
platformdirs==3.8.1
prometheus-client==0.17.1
prompt-toolkit==3.0.39
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.7.0
pylint==2.17.4
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
redis==4.6.0
requests==2.31.0
six==1.16.0
sqlparse==0.4.4
tomlkit==0.11.8
tornado==6.3.2
types-pytz==2023.3.0.0
types-PyYAML==6.0.12.10
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
wrapt==1.15.0
